requirements.txt
Flask==2.3.2
twilio==8.10.0
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.133.0
python-dotenv
google-genai
googletrans==4.0.0-rc1
google-generativeai
